factor out toolkit code
